Tie opinion expression and get my opinions with the proper user

+++ eos/smart-contracts/table_types.hpp
@@ -236,7 +236,7 @@ namespace conclubs {
     uint64_t user_id;
     uint64_t candidate_id;
     bool confidence;
-    uint32_t token_amount;
+    double token_amount;
 
     uint64_t primary_key() const { return id; };
     uint64_t get_user_id() const { return user_id; };
